relocate range components
