+++ pages/chapter4.tsx
@@ -1,7 +1,7 @@
 import Layout from '@/components/Layout';
 import PageCard from '@/components/PageCard';
 import JArrayReversion from '@/components/exercises/chapter-4-jon/JArrayReversion';
-import JSumRange from '@/components/exercises/chapter-4-jon/JSumRange';
+import JSumRange from '@/components/exercises/chapter-4-jon/range/JSumRange';
 import CityList from '@/components/exercises/chapter-4-oscar/AList/CityList';
 import DeepComparison from '@/components/exercises/chapter-4-oscar/DeepComparison/DeepComparison';
 import ReversingAnArray from '@/components/exercises/chapter-4-oscar/ReversingAnArray';
